完成属于Game的CookSystem及其测试

+++ Assets/Zephyr/GOAP/Sample/Game/Component/Count.cs
@@ -4,6 +4,6 @@ namespace Zephyr.GOAP.Sample.Game.Component
 {
     public struct Count : IComponentData
     {
-        public float Value;
+        public byte Value;
     }
 }
